refactor: add SQIDs encoding and standardise Long‑ID handling across publication and master layers
Implemented SQIDsEncoder in jooq query repository to obfuscate internal numeric IDs.
Re‑typed `PublicationRepository` and `AllowedFileTypeRepository` to use `Long` as primary key.
Adjusted services (`PublicationCommandService`, `MailArchiveCommandService`, `AllowedFileTypeService`) and mappers to accept `Long` IDs.
Reorganized DTO package hierarchy (documentType, mailCategory, mailType, quickMessage).
Added `PublicationDto` mapping logic using encoded IDs and optional totalCount handling.
Created comprehensive unit tests for `PublicationQueryService`.
Updated Gradle build to include Mockito dependencies and agent configuration for tests.

+++ src/main/java/id/perumdamts/mail/event/PublicationPublishedEvent.java
@@ -3,11 +3,11 @@ package id.perumdamts.mail.event;
 import java.time.Instant;
 
 public record PublicationPublishedEvent(
-        Integer publicationId,
+        Long publicationId,
         String publisherName,
         Instant publishedAt
 ) {
-    public PublicationPublishedEvent(Integer publicationId, String publisherName) {
+    public PublicationPublishedEvent(Long publicationId, String publisherName) {
         this(publicationId, publisherName, Instant.now());
     }
 }
